+++ src/main/java/com/example/crud_with_exception/doa/Studentdoa.java
@@ -3,6 +3,6 @@ package com.example.crud_with_exception.doa;
 import com.example.crud_with_exception.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Studentdoa extends JpaRepository<Student, Long> {
+public interface Studentdoa extends JpaRepository<Student, Integer> {
 
 }
